SE-300 better structure of the test folders
