+++ sdk/src/test/java/com/finbourne/features/dummyfiles/duplicates/DuplicateAnnotationsDummy.java
@@ -1,4 +1,4 @@
-package com.finbourne.features.unit.dummyfiles.duplicates;
+package com.finbourne.features.dummyfiles.duplicates;
 
 import com.finbourne.features.LusidFeature;
